chmodの権限でmysqlのバグ直す

+++ setup-query.sh
@@ -5,21 +5,26 @@ EDITED_FILE=./edited/mysqld.cnf
 sudo chmod 777 /var/log/mysql/
 sudo chmod 644 $TARGET_FILE
 if [ -e $TARGET_FILE ]; then
-sudo cp $TARGET_FILE $BACKUP_FILE
-sudo cp $TARGET_FILE $EDIT_FILE
-sudo chmod 777 $EDIT_FILE
-sudo chmod 777 $BACKUP_FILE
-if [ -e /var/log/mysql/mysql-slow.log ]; then
-rm -rf /var/log/mysql/mysql-slow.log
-fi
-sudo echo "slow_query_log = 1" >> $EDIT_FILE
-sudo echo "slow_query_log_file = /var/log/mysql/mysql-slow.log" >> $EDIT_FILE
-sudo echo "long_query_time = 0" >> $EDIT_FILE
+    if [ ! -e $EDIT_FILE ]; then
+        sudo cp $TARGET_FILE $BACKUP_FILE
+        sudo cp $TARGET_FILE $EDIT_FILE
+        sudo chmod 644 $EDIT_FILE
+        sudo chmod 644 $BACKUP_FILE
+    fi
+    if [ -e $EDIT_FILE ]; then
+        sudo cp $BACKUP_FILE $EDIT_FILE
+    fi
+    if [ -e /var/log/mysql/mysql-slow.log ]; then
+        rm -rf /var/log/mysql/mysql-slow.log
+    fi
+    sudo echo "slow_query_log = 1" >> $EDIT_FILE
+    sudo echo "slow_query_log_file = /var/log/mysql/mysql-slow.log" >> $EDIT_FILE
+    sudo echo "long_query_time = 0" >> $EDIT_FILE
 
-sudo cp $EDIT_FILE $TARGET_FILE
-sudo cp $EDIT_FILE $EDITED_FILE
-sudo chmod 777 $EDITED_FILE
-sudo rm -rf $EDIT_FILE
+    sudo cp $EDIT_FILE $TARGET_FILE
+    sudo cp $EDIT_FILE $EDITED_FILE
+    sudo chmod 644 $EDITED_FILE
+    sudo rm -rf $EDIT_FILE
 fi
 
 if [ ! -e $TARGET_FILE ]; then
